admin's book options add

+++ java/src/main/java/xzz/library/service/serviceImpl/AdminServiceImpl.java
@@ -45,6 +45,8 @@ public class AdminServiceImpl implements AdminService {
     @Override
     @Transactional
     public String addUser(User user) {
+        if (user == null)
+            return "用户错误";
         if (userMapper.getUserByUsername(user.getUsername()) != null){
             return "用户名重复";
         }
@@ -61,6 +63,8 @@ public class AdminServiceImpl implements AdminService {
     @Override
     @Transactional
     public String deleteUser(String[] ids) {
+        if (ids == null)
+            return "用户错误";
         try {
             for (String id : ids)
                 userMapper.deleteByPrimaryKey(id);
@@ -74,12 +78,81 @@ public class AdminServiceImpl implements AdminService {
     @Override
     @Transactional
     public String updateUser(User user) {
+        if (user == null)
+            return "用户错误";
         try {
             userMapper.updateByPrimaryKey(user);
             return null;
         }catch (Exception e){
             e.printStackTrace();
             return "修改失败";
+        }
+    }
+
+    //========================================================================================
+    //Books
+    //========================================================================================
+    @Override
+    public BooksDto getBooks(Book book, Integer limit, Integer page) {
+        if (limit == null)
+            limit = 10;
+        int start = 0;
+        if (page != null) {
+            start = (page-1)*limit;
+        }
+        if (book == null)
+            book = new Book();
+        List<Book> books = bookMapper.getBookList(book.getBookName(), book.getBookType(), book.getAuthor(),
+                book.getPublisher(),book.getStock(), limit, start);
+        int total = bookMapper.countBook(book.getBookName(), book.getBookType(), book.getAuthor(),
+                book.getPublisher(),book.getStock());
+        return new BooksDto(books,total);
+    }
+
+    @Override
+    @Transactional
+    public String addBook(Book book) {
+        if (book == null)
+            return "图书错误";
+        book.initial();
+        try {
+            bookMapper.insert(book);
+            return null;
+        }catch (Exception e){
+            e.printStackTrace();
+            return "图书添加失败";
+        }
+    }
+
+    @Override
+    public Book getBook(String id) {
+        return bookMapper.selectByPrimaryKey(id);
+    }
+
+    @Override
+    @Transactional
+    public String deleteBook(String[] ids) {
+        if (ids == null)
+            return "图书错误";
+        try {
+            for (String id : ids)
+                bookMapper.deleteByPrimaryKey(id);
+            return null;
+        } catch (Exception e){
+            e.printStackTrace();
+            return "删除图书失败";
+        }
+    }
+
+    @Override
+    public String updateBook(Book book) {
+        if(book == null)
+            return "图书错误";
+        try{
+            bookMapper.updateByPrimaryKey(book);
+        }catch (Exception e){
+            e.printStackTrace();
+            return "修改图书失败";
         }
     }
